+++ resources/manage_puma.rb
@@ -1,5 +1,5 @@
 #
-# Author:: Earth U (<iskitingbords @ gmail.com>)
+# Author:: Earth U (<iskitingbords@gmail.com>)
 # Cookbook Name:: app-ror
 # Resource:: manage_puma
 #
@@ -18,54 +18,64 @@
 # limitations under the License.
 #
 
-# Add config for Jungle, the Puma manager.
-# (https://github.com/puma/puma/tree/master/tools/jungle/upstart)
+# Add config for Puma manager.
+# (Based on: https://github.com/puma/puma/tree/master/tools/jungle/upstart)
 
 # Project properties
-property :roots, [Array, String], name_property: true
-property :roots_conf_path, String, default: '/etc/puma.conf'
-property :project_conf_path, String, default: '../../shared/puma.rb'
+property :app_dir, String, name_property: true
+property :conf_path, String, default: '../../shared/puma.rb'
 
 # App properties
 property :user, String, required: true
 property :group, [String, false], default: false
+property :env_file, [String, false], default: false
 
-property :puma_name, String, default: 'puma.conf'
-property :puma_source, String, default: 'puma.conf.erb'
+property :puma_source, [String, false], default: false
 property :puma_cookbook, String, default: 'app-ror'
-property :puma_manager_name, String, default: 'puma-manager.conf'
-property :puma_manager_source, String, default: 'puma-manager.conf.erb'
-property :puma_manager_cookbook, String, default: 'app-ror'
 
 action :install do
-
-  template "/etc/init/#{new_resource.puma_manager_name}" do
-    source new_resource.puma_manager_source
-    cookbook new_resource.puma_manager_cookbook
-    variables({ :roots_conf_path => new_resource.roots_conf_path })
+  if not node['platform'] == 'ubuntu'
+    Chef::Application.fatal!("#{node['platform']} is not supported")
   end
 
-  group = new_resource.group ? new_resource.group : new_resource.user
-  template "/etc/init/#{new_resource.puma_name}" do
-    source new_resource.puma_source
-    cookbook new_resource.puma_cookbook
-    variables(
-      :user => new_resource.user,
-      :group => group,
-      :project_conf_path => new_resource.project_conf_path
-    )
-  end
+  env_file = new_resource.env_file ? new_resource.env_file : "/home/#{new_resource.user}/.etc/ruby_env"
 
-  roots = if new_resource.roots.is_a?(String)
-    new_resource.roots.split(',')
+  if node['platform_version'].to_f >= 15.04
+
+    puma_source = new_resource.puma_source ? new_resource.puma_source : 'puma.service.erb'
+    template '/etc/systemd/system/puma.service' do
+      source   puma_source
+      cookbook new_resource.puma_cookbook
+      variables(
+        :user      => new_resource.user,
+        :env_file  => env_file,
+        :app_dir   => new_resource.app_dir,
+        :conf_path => new_resource.conf_path
+      )
+      notifies :run, 'execute[enable_puma_service]', :immediately
+    end
+
+    execute 'enable_puma_service' do
+      command 'systemctl enable puma.service'
+      action  :nothing
+    end
+
+  elsif node['platform_version'].to_f >= 12.04
+
+    puma_source = new_resource.puma_source ? new_resource.puma_source : 'puma.conf.erb'
+    group = new_resource.group ? new_resource.group : new_resource.user
+    template '/etc/init/puma.conf' do
+      source   puma_source
+      cookbook new_resource.puma_cookbook
+      variables(
+        :user      => new_resource.user,
+        :group     => group,
+        :env_file  => env_file,
+        :app_dir   => new_resource.app_dir,
+        :conf_path => new_resource.conf_path
+      )
+    end
   else
-    new_resource.roots
+    Chef::Application.fatal!("Version #{node['platform_version']} is not supported")
   end
-
-  template new_resource.roots_conf_path do
-    source 'puma-roots.conf.erb'
-    cookbook 'app-ror'
-    variables({ :roots => roots })
-  end
-
 end