+++ test/cookbooks/test/metadata.rb
@@ -1,0 +1,9 @@
+name             'test'
+maintainer       'Earth U'
+maintainer_email 'iskitingbords@gmail.com'
+license          'Apache License'
+description      'Installs/Configures test'
+long_description 'Installs/Configures test'
+version          '0.1.0'
+
+depends 'app-ror'