+++ test/cookbooks/test/recipes/manage_puma.rb
@@ -1,0 +1,23 @@
+#
+# Author:: Earth U (<iskitingbords@gmail.com>)
+# Cookbook Name:: test
+# Recipe:: manage_puma
+#
+# Copyright (C) 2018, Earth U
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+app_ror_manage_puma "#{node['test']['base_dir']}/current" do
+  user node['test']['user']
+end