+++ test/integration/default/base_dirs.rb
@@ -1,0 +1,23 @@
+describe directory('/var/src/test') do
+  it { should exist }
+  its('owner') { should eq 'ubuntu' }
+  its('group') { should eq 'ubuntu' }
+end
+
+describe directory('/var/src/test/shared') do
+  it { should exist }
+  its('owner') { should eq 'ubuntu' }
+  its('group') { should eq 'ubuntu' }
+end
+
+describe directory('/var/src/test/shared/config') do
+  it { should exist }
+  its('owner') { should eq 'ubuntu' }
+  its('group') { should eq 'ubuntu' }
+end
+
+describe directory('/var/src/test/tmp') do
+  it { should exist }
+  its('owner') { should eq 'ubuntu' }
+  its('group') { should eq 'ubuntu' }
+end